p4i integration Integrate 262521 from 4.3 to main: Fixes:    Remove all compiler warnings on MinGW, except for the 3rd party libraries. Task:     142171 RevBy:    Prasanth/(Simon)
[git-p4: depot-paths = "//depot/qt/main/": change = 262522]

+++ src/corelib/io/qfilesystemwatcher_win.cpp
@@ -70,7 +70,7 @@ void QWindowsFileSystemWatcherEngine::run()
                     qDebug("QWindowsFileSystemWatcherEngine: unknown message '%c' send to thread", char(m));
                 }
                 break;
-            } else if (r > WAIT_OBJECT_0 && r < WAIT_OBJECT_0+handlesCopy.count()) {
+            } else if (r > WAIT_OBJECT_0 && r < WAIT_OBJECT_0 + uint(handlesCopy.count())) {
                 int at = r - WAIT_OBJECT_0;
                 Q_ASSERT(at < handlesCopy.count());
                 HANDLE handle = handlesCopy.at(at);
